break setup and finders out from Table into Queryable
