+++ lib/table.js
@@ -3,144 +3,21 @@ var assert = require("assert");
 var clone = require('clone');
 var util = require('util');
 var DocumentTable = require("./document_table");
+var Queryable = require("./queryable");
 var Where = require("./where");
 var ArgTypes = require("./arg_types");
 var DA = require("deasync");
 
 //a simple wrapper for a table
-var Table = function(args){
-  this.schema = args.schema;
-  this.name = args.name;
+var Table = function(args) {
+  Queryable.apply(this, arguments);
+
   this.pk = args.pk;
-  this.db = args.db;
 
-  // Build a fully qualified name if the schema is other than public:
-  this.fullname = this.name;
-  if(this.schema !== "public") {
-    this.fullname = this.schema + "." + this.name;
-  }
-
-  // build delimited names in one spot instead of when building sql:
-  this.delimitedName = "\"" + this.name + "\"";
-  this.delimitedSchema = "\"" + this.schema + "\"";
-  this.delimitedFullName = this.delimitedName;
-
-  // if the schema is other than public blah blah...
-  if(this.schema !== "public") {
-    this.delimitedFullName = this.delimitedSchema + "." + this.delimitedName;
-  }
   _.extend(this,DocumentTable);
 };
 
-//a simple alias for returning a single record
-Table.prototype.findOne = function(args, next){
-  if(_.isFunction(args)){
-    next = args;
-    args = {};
-  }
-
-  this.find(args, function(err,results){
-    if(err){
-      next(err,null);
-    }else{
-      var result;
-
-      if (_.isArray(results)) {
-        if (results.length > 0) { result = results[0]; }
-      } else {
-        result = results;
-      }
-
-      next(null,result);
-    }
-  });
-};
-Table.prototype.findOneSync = DA(Table.prototype.findOne);
-
-/**
- * Counts rows and calls back with any error and the total. There are two ways to use this method:
- *
- * 1. find() style: db.mytable.count({field: value}, callback);
- * 2. where() style: db.mytable.count("field=$1", [value], callback);
- */
-Table.prototype.count = function() {
-  var args;
-  var where;
-
-  if (_.isObject(arguments[0])) {
-    args = ArgTypes.findArgs(arguments);
-    where = _.isEmpty(args.conditions) ? {where : " "} : Where.forTable(args.conditions);
-  } else {
-    args = ArgTypes.whereArgs(arguments);
-    where = {where: " where " + args.where};
-  }
-
-  var sql = "select COUNT(1) from " + this.delimitedFullName + where.where;
-
-  this.db.query(sql, where.params || args.params, {single : true}, function(err, res) {
-    if (err) args.next(err, null);
-    else args.next(null, res.count);
-  });
-};
-Table.prototype.countSync = DA(Table.prototype.count);
-
-//a simple way to just run something
-//just pass in "id=$1" and the criteria
-Table.prototype.where = function(){
-  var args = ArgTypes.whereArgs(arguments);
-
-  var sql = "select * from " + this.delimitedFullName + " where " + args.where;
-  this.db.query(sql, args.params, args.next);
-};
-Table.prototype.whereSync = DA(Table.prototype.where);
-
-Table.prototype.find = function(){
-  var args = ArgTypes.findArgs(arguments);
-
-  //set default options
-  args.options.order || (args.options.order = util.format('"%s"', this.pk));
-  args.options.limit || (args.options.limit = "1000");
-  args.options.offset || (args.options.offset = "0");
-  args.options.columns || (args.options.columns = "*");
-
-  if(_.isFunction(args.conditions)){
-    //this is our callback as the only argument, caught by Args.ANY
-    args.next = args.conditions;
-  }
-
-  var returnSingle = false;
-  var where, order, limit, cols="*", offset;
-
-  if(args.options.columns){
-    if(_.isArray(args.options.columns)){
-      cols = args.options.columns.join(',');
-    }else{
-      cols = args.options.columns;
-    }
-  }
-  order = " order by " + args.options.order;
-  limit = " limit " + args.options.limit;
-  offset = " offset " + args.options.offset;
-
-  if(_.isNumber(args.conditions)){
-    //a primary key search
-    var newArgs = {};
-    newArgs[this.primaryKeyName()] = args.conditions;
-    args.conditions = newArgs;
-    returnSingle = true;
-  }
-
-  where = _.isEmpty(args.conditions) ? {where : " "} : Where.forTable(args.conditions);
-
-  var sql = "select " + cols + " from " + this.delimitedFullName + where.where + order + limit + offset;
-
-  if (args.options.stream) {
-    this.db.stream(sql, where.params, null, args.next);
-  } else {
-    this.db.query(sql, where.params, {single : returnSingle}, args.next);
-  }
-};
-Table.prototype.findSync = DA(Table.prototype.find);
+util.inherits(Table, Queryable);
 
 Table.prototype.insert = function(data, next) {
   if(!data) throw "insert should be called with data";
